% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gap_filling.R
\name{SAXGAP}
\alias{SAXGAP}
\title{Fill MS-gaps using SAX}
\usage{
SAXGAP(
  peak_table,
  mzML_filenames,
  SAXGAP_params = list(a = 6, w = 6, t = 14, pe = 0.9, intensity_threshold = 10000,
    dppm = 5),
  output_file
)
}
\arguments{
\item{peak_table}{A mandatory character vector giving the path to an aligned peak table in csv format or a data.table that is the aligned peak table. See the Details section for more information on how this parameter and the filenames parameter interact.}

\item{mzML_filenames}{The names of the .mzML-mzML_filenames that are references in peak table. See the Details section for more information}

\item{SAXGAP_params}{A list containing the main parameters for the SAXGAP function. See the Details section for more information}

\item{output_file}{An optional charcter vector giving the filepath of where the output table should be stored.}
}
\value{
The function returns the gap filled table with the gap filled peaks stored as a negative value and an added \code{Removal_Candidate} column that signifies whether a peak was likely picked in error. 
If output_file is set it writes the same table to the path specified by output_file.
}
\description{
Function that fills gaps utilizing SAX (Symbol Aggregate approXimation)
}
\details{
\subsection{peak_table and mzML_filename:}{

The peak_table and mzML_filenames parameter need to have the same underlying names for the storage of the relationship of the intensities and retention times to the files. 
If, for example, one supplies the file \code{Sample_1.mzML} then the peak_table MUST contain a column named either \code{Intensity_Sample_1} or \code{Intensity_Sample_1.mzML}. The holds true for
retention time, i.e. the table needs to have a column named either \code{RT_Sample_1.mzML} or \code{RT_Sample_1}. The peak table also needs general \code{mz} and \code{RT} columns. All retention times must provided in seconds!

}

\subsection{SAXGAP_params:}{
The SAXGAP_params paarmeter must be a named list containing the parameters:
\describe{
 \item{a}{The alphabet size of the SAX conversion and comparison}
 \item{w}{The word length for peaks}
 \item{t}{The length of the extracted EICs}
 \item{pe}{The percentage cutoff for peak removal}
 \item{dppm}{The instrument-specific allowed ppm error of the peaks}
 \item{intensity_threshold}{The minimum intensity for peaks}
}
}
}
